pyproject.toml
[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
show_error_codes = true

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "massconfigmerger"
version = "0.4.0"
description = "VPN subscription aggregator and merger"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp",
    "aiodns",
    "nest-asyncio",
    "telethon",
    "PyYAML",
    "geoip2",
    "pydantic>=2",
    "pydantic-settings",
    "tqdm",
    "flask",
    "aiosqlite",
    "prometheus_client"
]

[project.scripts]
massconfigmerger = "massconfigmerger.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pyfakefs",
    "types-PyYAML",
    "types-requests",
    "flake8",
    "mypy",
    "pre-commit",
    "pip-audit",
]
web = [
    "Flask",
    "asgiref"
]

[tool.setuptools.packages.find]
where = ["src"]
